+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/repositories/MedicalHistoryRepo.java
@@ -8,10 +8,10 @@ import java.util.Date;
 public interface MedicalHistoryRepo extends JpaRepository<MedicalHistory, Integer> {
 
     // Fetch all medical history records for a given patient
-    List<MedicalHistory> findAllByPatient_id(int patient_id);
+    List<MedicalHistory> findAllByPatient_id(Long patient_id);
 
     // Fetch medical history by doctor
-    List<MedicalHistory> findAllByDoctor_id(int doctor_id);
+    List<MedicalHistory> findAllByDoctor_id(Long doctor_id);
 
     // Get active medical histories (no end_date yet)
     List<MedicalHistory> findAllByEnd_dateIsNull();
