+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/repositories/MedicationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MedicationRepository extends JpaRepository<Medication, Integer> {
-    List<Medication> findAllByPatient_id(int patient_id);
+    List<Medication> findAllByPatient_id(Long patient_id);
 
 }
